Change to f32 and add CLI arg parsing

+++ src/camera.rs
@@ -1,14 +1,14 @@
 use crate::geometry::{Point, Ray, Vec3};
 
 pub struct Camera {
-    origin: Point<f64>,
-    lower_left_corner: Point<f64>,
-    vertical: Vec3<f64>,
-    horizontal: Vec3<f64>,
+    origin: Point<f32>,
+    lower_left_corner: Point<f32>,
+    vertical: Vec3<f32>,
+    horizontal: Vec3<f32>,
 }
 
 impl Camera {
-    pub fn new(viewport_height: f64, viewport_width: f64, focal_length: f64) -> Self {
+    pub fn new(viewport_height: f32, viewport_width: f32, focal_length: f32) -> Self {
         let origin = Point::new(0.0, 0.0, 0.0);
         let horizontal = Vec3::new(viewport_width, 0.0, 0.0);
         let vertical = Vec3::new(0.0, viewport_height, 0.0);
@@ -25,7 +25,7 @@ impl Camera {
         }
     }
 
-    pub fn ray_at(&self, u: f64, v: f64) -> Ray {
+    pub fn ray_at(&self, u: f32, v: f32) -> Ray {
         Ray::new(
             self.origin,
             self.horizontal * u + self.vertical * v + self.lower_left_corner.into(),
