+++ src/geometry/ray.rs
@@ -1,25 +1,25 @@
-use super::{Double, Point, Vec3};
+use super::{Point, Vec3};
 
 #[derive(Debug, Copy, Clone)]
 pub struct Ray {
-    origin: Point<Double>,
-    direction: Vec3<Double>,
+    origin: Point<f32>,
+    direction: Vec3<f32>,
 }
 
 impl Ray {
-    pub fn new(origin: Point<Double>, direction: Vec3<Double>) -> Ray {
+    pub fn new(origin: Point<f32>, direction: Vec3<f32>) -> Ray {
         Ray { origin, direction }
     }
 
-    pub fn origin(self) -> Point<Double> {
+    pub fn origin(self) -> Point<f32> {
         self.origin
     }
 
-    pub fn direction(self) -> Vec3<Double> {
+    pub fn direction(self) -> Vec3<f32> {
         self.direction
     }
 
-    pub fn at(self, t: Double) -> Point<Double> {
+    pub fn at(self, t: f32) -> Point<f32> {
         self.origin + (self.direction * t).into()
     }
 }
